#Card Logic
Created basic logic to split deck in half and assign to both players ibject property. This is done by manipulating a copy of the main decks array. I wrote a function to shuffle the array but I ran into a problem of the result getting infinetely nested..

WIll come back to the shuffle feature when games foundation has been more established.